fix(models): Float→Numeric(12,2) for monetary fields in response.py

+++ app/models/response.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, DateTime, Enum, ForeignKey, Text
+from sqlalchemy import Column, Integer, Float, DateTime, Enum, ForeignKey, Text, Numeric
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 import enum
@@ -17,9 +17,9 @@ class Response(Base):
     load_id   = Column(Integer, ForeignKey("loads.id"), nullable=False)
     user_id   = Column(Integer, ForeignKey("users.id"), nullable=False)
     message   = Column(Text, nullable=True)
-    price_usd = Column(Float, nullable=True)    # предложенная цена в USD
-    price_gel = Column(Float, nullable=True)    # предложенная цена в GEL
-    exchange_rate_at_creation = Column(Float, nullable=True)  # GEL/USD из NBG
+    price_usd = Column(Numeric(12, 2), nullable=True)    # предложенная цена в USD
+    price_gel = Column(Numeric(12, 2), nullable=True)    # предложенная цена в GEL
+    exchange_rate_at_creation = Column(Numeric(12, 6), nullable=True)  # GEL/USD из NBG
     status    = Column(Enum(ResponseStatus), default=ResponseStatus.pending)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
